backend/requirements.txt
azure-ai-inference==1.0.0b9
azure-core==1.34.0
certifi==2025.4.26
charset-normalizer==3.4.2
idna==3.10
isodate==0.7.2
python-dotenv==1.1.0
requests==2.32.4
six==1.17.0
typing_extensions==4.14.0
urllib3==2.4.0
